+++ frontend/webapp/containers/overview/destination/destination.tsx
@@ -1,50 +1,64 @@
 "use client";
 import React, { useState } from "react";
 import { KeyvalLoader } from "@/design.system";
-import { OVERVIEW, QUERIES } from "@/utils/constants";
+import { NOTIFICATION, OVERVIEW, QUERIES } from "@/utils/constants";
 import { useQuery } from "react-query";
 import { getDestinations } from "@/services";
 import { OverviewHeader, DestinationsManagedList } from "@/components/overview";
 import { DestinationContainerWrapper } from "./destination.styled";
 import { NewDestinationFlow } from "./new.destination.flow";
 import { UpdateDestinationFlow } from "./update.destination.flow";
+import { useNotification } from "@/hooks";
 
 export function DestinationContainer() {
   const [selectedDestination, setSelectedDestination] = useState<any>(null);
   const [displayNewDestination, setDisplayNewDestination] =
     useState<boolean>(false);
-
+  const { show, Notification } = useNotification();
   const {
     isLoading: destinationLoading,
     data: destinationList,
     refetch,
   } = useQuery([QUERIES.API_DESTINATIONS], getDestinations);
 
-  function handleAddNewDestinationClick() {
-    setDisplayNewDestination(true);
+  function onSuccess(message = OVERVIEW.DESTINATION_UPDATE_SUCCESS) {
+    refetch();
+    setSelectedDestination(null);
+    setDisplayNewDestination(false);
+    show({
+      type: NOTIFICATION.SUCCESS,
+      message,
+    });
+  }
+
+  function onError({ response }) {
+    const message = response?.data?.message;
+    show({
+      type: NOTIFICATION.ERROR,
+      message,
+    });
   }
 
   if (destinationLoading) {
     return <KeyvalLoader />;
   }
 
-  if (displayNewDestination) {
-    return (
-      <NewDestinationFlow
-        onBackClick={() => {
-          refetch();
-          setDisplayNewDestination(false);
-        }}
-      />
-    );
-  }
-
   return (
     <DestinationContainerWrapper>
-      {selectedDestination ? (
+      {displayNewDestination ? (
+        <NewDestinationFlow
+          onSuccess={onSuccess}
+          onError={onError}
+          onBackClick={() => {
+            setDisplayNewDestination(false);
+          }}
+        />
+      ) : selectedDestination ? (
         <UpdateDestinationFlow
           selectedDestination={selectedDestination}
           setSelectedDestination={setSelectedDestination}
+          onSuccess={onSuccess}
+          onError={onError}
         />
       ) : (
         <>
@@ -52,10 +66,11 @@ export function DestinationContainer() {
           <DestinationsManagedList
             data={destinationList}
             onItemClick={setSelectedDestination}
-            onMenuButtonClick={handleAddNewDestinationClick}
+            onMenuButtonClick={() => setDisplayNewDestination(true)}
           />
         </>
       )}
+      <Notification />
     </DestinationContainerWrapper>
   );
 }
